refactor: add internship component
